Adds prismjs and Planning to Learn article
Wanted to have the code highlighting locally so to make it easier to digest
I have added prismjs. Still need to create a theme for it. With that in
I updated the two articles that used syntax highlighting and removed a
now unused plugin.

Added Planning to Learn article. Will need to look into twitter embedding
and see if there's a more performant way of doing that. It has meant adding
another plugin to handle that