Converted page to TypeScript, improved semantic meaning
Missed one of the pages when I did the TypeScript conversion, so the blog
index is now also TypeScript and simplified.

Converted the blog date to a time element to improve the semantic meaning.
Added styles to the blog roll items to help that.

Made the post feature image a link into the blog post to make it easier
to access blog posts