Adds article about git commit messsages
Feel like I have to make a really good git commit now that I have written
this. What pressure!

Had several discussions with people at work about the length of commit
messages, so wanted to put together a quick explainer as to why I structure
them the way that I do, which is to have history kept within git instead
of the ui that the PR is opened in.